everything changed to uint64_t

+++ Includes/ConfigGenerator/ConfigGenerator.hpp
@@ -8,7 +8,7 @@
 class ConfigGenerator
 {
 	public:
-		ConfigGenerator(size_t kmer_len, uint64_t w_min, uint64_t w_max, uint32_t n, uint64_t mask,
+		ConfigGenerator(uint64_t kmer_len, uint64_t w_min, uint64_t w_max, uint32_t n, uint64_t mask,
 				uint64_t number_of_samples, uint64_t seq_len, std::string data_generator, std::string data_file_name);
 		void add_data_generator_config(std::string output_path);
 		void add_hasher_config(std::string output_path, Json::Value config);
@@ -17,7 +17,7 @@ class ConfigGenerator
 		void add_number_of_samples_config(std::string output_path, Json::Value config);
 
 	protected:
-		size_t kmer_len;
+		uint64_t kmer_len;
 		uint64_t w_min;
 		uint64_t w_max;
 		uint32_t n;
