+++ Includes/SeedCreators/RandStrobeCreators/RandStrobeCreatorSpecialCaseFixedShen.hpp
@@ -6,8 +6,8 @@
 class RandStrobeCreatorSpecialCaseFixedShen : public RandStrobeCreatorShen
 {
 	public:
-		RandStrobeCreatorSpecialCaseFixedShen(Hasher* hasher, Comparator* comparator, size_t kmer_len, size_t w_min, size_t w_max,
-				uint32_t n = DEFAULT_N, uint64_t mask = DEFAULT_MASK, uint64_t q = DEFAULT_Q);
+		RandStrobeCreatorSpecialCaseFixedShen(Hasher* hasher, Comparator* comparator, uint64_t kmer_len, uint64_t w_min, uint64_t w_max,
+				uint64_t n = DEFAULT_N, uint64_t mask = DEFAULT_MASK, uint64_t q = DEFAULT_Q);
 	protected:
 		void prepare_data();
 		uint64_t get_new_curr_hash(const Strobemer* strobemer);
